requirements.txt
asgiref==3.8.1
aura-sr==0.0.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.6
einops==0.8.1
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
